+++ includes/admin/class-admin.php
@@ -89,6 +89,7 @@ if ( ! class_exists( 'um\admin\Admin' ) ) {
 			$this->enqueue();
 			$this->notices();
 			$this->secure();
+			$this->site_health();
 		}
 
 		function init_variables() {
@@ -2076,6 +2077,18 @@ if ( ! class_exists( 'um\admin\Admin' ) ) {
 				UM()->classes['um\admin\secure'] = new Secure();
 			}
 			return UM()->classes['um\admin\secure'];
+		}
+
+		/**
+		 * @since 2.6.12
+		 *
+		 * @return Site_Health
+		 */
+		public function site_health() {
+			if ( empty( UM()->classes['um\admin\site_health'] ) ) {
+				UM()->classes['um\admin\site_health'] = new Site_Health();
+			}
+			return UM()->classes['um\admin\site_health'];
 		}
 	}
 }
